+++ lib/v2/abmedia/router.js
@@ -1,3 +1,4 @@
 module.exports = function (router) {
-    router.get('/:category?', require('./index'));
+    router.get('/index', require('./index'));
+    router.get('/:category?', require('./category'));
 };
